stockalert/pyproject.toml
[tool.poetry]
name = "stockalert"
version = "0.1.0"
description = "Modular stock alert backend with FastAPI, Postgres, WebSockets, and provider abstraction"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}, {include = "tests"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = ">=0.111"
uvicorn = {extras = ["standard"], version = ">=0.29"}
alpaca-py = ">=0.25"
sqlalchemy = {extras = ["asyncio"], version = ">=2.0"}
asyncpg = ">=0.29"
pydantic = ">=2.6"
pandas = ">=2.2"
numpy = ">=1.26"
python-dotenv = ">=1.0"
aiohttp = ">=3.9"
alembic = "^1.16.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.21"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
